Update synthwave.yaml
I am just putting in suggestions to fix the outstanding issues with the theme. I have made the changes locally and it looks good. 
If you would like to see some more of the colour changes I have made I can send you a copy of my yaml. but great theme and I am definitely a child of the 80's and love the sythwave vibe.